🔥Quest: queue&createQueue (#171)
First half of the commit is the removal of the createQuest and the corresponding modifications to the test to get them back to passing, then I pulled the actual queue function off of Quest.sol.

As it stands now, I just set the queued bool in the initialize and fire off the Queued event right in that function, but we could leave queued false and remove the event pretty easily assuming nothing is listening for, or checking that variable.

Here's the main logic change in QuestFactory to accommodate the lack of queue and here's where I removed the actual queue function. Everything else is pretty straight forward.